# ws2812 strip data buffer