Following #790, here is my current work almost ready to merge.

Feel free to comment or ask if something seems missing.

Goal was to efficiently store a RW state buffer of the led strip. At each iteration loop it is then easier to update only a portion of it, or compute the step in multiple calls without doing insane string concatenations.

Led chaser example:
```lua
local i, dir, buf = 0, 1, ws2812.newBuffer(300); buf:fill(0, 0, 0); tmr.alarm(0, 15, 1, function()
    i=i+dir
    if i == 299 then
      dir = -1
    elseif i == 0 then
      dir = 1
    end
    buf:fade(2)
    buf:set(i, 0, 255, 0)
    buf:write(4)
end)
```

Currently, :set allows (G, R, B) tuple, { G, R B } table and "GRBGRB..." string.
In the future, I plan to add table of strings { "GRB", "GRB", "GRB", ... } and table of tables { { G, R B }, { G, R B }, { G, R B }, ... } to be able to store precomputed patterns.