@TerryE 2016-01-05
This is a general comment about all bit banging drivers but including this one. IMO, the general approach is unstable with the ESP SDK. The general guidelines are that interrupt handlers and disabling of interrupts should be no more than 50uS and any cb task less than 50-100 msec. We've discussed this second but not the first. You are doing a `ets_intr_lock` ...`ets_intr_unlock` around your entire routine, and if you do this there is no guarantee that anything that depends on interrupts will still be working at the end of it. OK, the bit timing within the 24 bit triplet is tight,  but this is only ~27 uS. The reset on most real devices seems to be around 50 mSec, so why not limit your lock to around the 24 bit sequence so that any pending interrupts can be delivered between 3 byte sets? The library will then be a lot more stable.

.