@Alkorin 2016-01-05
Reset timing is 50us, not 50ms, so I should have the guarantee that the SDK will not interrupt me more than 50us.

I'll give it a try but I have already seen +/- 1ms deviation on some timing benchmarks due to interrupts :/