@TerryE 2016-01-05
You are working within a set PC ground rules set by Espressif and their SDK. You are free to ignore them for your own code, but a Lua developer using your module might expect the Wifi to work for example with this driver and start raising issues if it doesn't.

That 1ms pf other