@Alkorin 2016-01-05
Yes I'll check it this evening @TerryE 

I need wifi/network too, so I'll take care of it ;)