@TerryE 2016-01-05
I realise that the datasheet says that the WS2812 reset interval is > 50µS but it doesn't say how much.  I was looking at a couple of detailed posts which retro-engineer the operational characteristics and they are very inconsistent. (For example [this one](http://wp.josh.com/2014/05/13/ws2812-neopixels-are-not-so-finicky-once-you-get-to-know-them/) suggests 50 mSec!)  As fas a I can see, the sequence goes **<reset>(<RGB>)\*** where the reset is passed down the line and the first chip pops and latches the first **RGB** and passes the remaining (?low?) + (n-1) **RGB** sequence to the next chip and so on.  If this is correct then the tail-end Charlie on a 1000 long strip will see <reset+8991µS low>**RGB**.

So if you are correct in asserting the <reset> **RBG RBG RBG** <interrupt delay>** RGB . . . ** will be seen as a <reset> 3×RGB <reset> 997×RGB then we can't allow interrupts in the ~9mSec window when we are bit banging.  But in this case we should also assume that the Wifi stack _might_ have died or hung in the meantime and therefore the programming guidelines for the driver should explain this constraint end give an application example of how the Lua programmer can reset and recover the Wifi.

Such are the joys of a single core CPU and a closed source HW architecture + SDK. :cry: