@TerryE 2016-01-06
Just brooding about this a bit more. The time for a single RGB is 27 uSec. So we can clearly disable interrupts for this. But even a stop of 10, let alone 100, LEDs gives an end to end interval that is to long to disable interrupts safely so I suggest that the obvious option is to enable interrupts between RGB codes. So we really have three options:

  *  Ignore the SDK timing guideline, disable interrupts for the entire RGB sequence, and then try to repair the damage when done (e.g. restart the WiFi and TCP stacks).

  *  Enable interrupts between RBG sets, but ignore the potential consequences of an interrupt having been delivered.

*  Enable interrupts between RBG sets, and add detection logic around the disable..enable to see if an interrupt has been delivered and have some recovery rules -- for example if the delay was more than X uSec, then reset and restart the sequence.

This last option might be the way to go if the probability of a disruptive interrupt is small and the visual impact of the restart isn't perceivable.