@nickandrew 2016-01-06
A technique I have used very successfully on the _ATTiny85_ is that an interrupt handler does the actual bit banging (controlled by a timer, for precision) under the direction of non-interrupt code.

To describe it informally, the mainline code would start an operation (such as sending a character down the software emulated serial port) and then setup the 2nd operation (the next character to be sent) and then wait until the first operation had finished, then setup the 3rd operation, and so forth. So the interrupt code would always have 1 buffered character to send, and it would be able to start on the next character immediately after finishing the previous. There'd be no delay proceeding from one operation to the next.