@Alkorin 2016-01-06
Another solution for my own project would be to add another µC to do the timing critical stuff, but it's not the question here.

I'll check interrupts durations to see if we could let them enabled or not. If not, we should let a note in the documentation which says that the wifi stack may be broken if the led strip is too long.

What are the guidelines for the maximum time that interrupts can be disabled ?