@karrots 2016-01-06
Not sure if this will help things. This guy has an extensive write up on the timing needs of the WS2812 protocol

https://cpldcpu.wordpress.com/2014/01/14/light_ws2812-library-v2-0-part-i-understanding-the-ws2812/