@Alkorin 2016-01-06
Thanks a lot @karrots , it looks like a 1 could be a lot longer, so ... we could try to deliver interrupts while we try to send a 1 :)

0 is a "short pulse", 1 is a "long pulse". If we send a "0", it's timing sensitive. If we send a 1 ... we maybe have a winner there !, Will try this tomorrow !!!