@Alkorin 2016-01-07
Hum, no :/
Total cycle time can't exceed 9µs, so if an interrupt occurs, it should not be more than 9µs