@TerryE 2016-01-08
@Alkorin Thomas, reread my last comment and consider my option(3).  Let's say for the sake of argument you are getting 10 interrupts per second on you  ws2812 chip.  We should treat this Poisson but let's keep this simple an assume that this is 1 every 100 mSec or so.  If you are driving a 100 LED strip then the time to relatch the RGB data is 2.7mSec, so if you are enabling interrupts every 27µSec with collision detection then roughly 1 in 40 RGB loads will have a detected collision.  Let's say this occurs after the 16 LED load. What is wrong with just letting the sequence reset and then restarting the sequence. So instead of:

*    <reset>100×RBG

You do

*    <reset>16×RBG (interrupt detected) <reset>100×RBG