@Alkorin 2016-01-10
Yes, it works @TerryE, but as expected it resets :)

And it resets fast ... even at 160MHz where interrupts are handled in 6.5us, I've 2 or 3 currupted send per seconds. I'll implement the interrupt detection and see.

Do we have some "known" values for the maximum time that interrupts can be disabled ?