@TerryE 2016-01-10
non-OS V1.5 SDK Programming Guide Section 2 -- Overview

> Notice:
>
>  *  Using non-OS SDK which is single-threaded, the CPU should not take long to execute tasks:
>     *   If interrupt is disabled, CPU can only be occupied in us range and the time should not bemore than 10 µs; if interrupt is not disabled, it is suggested that CPU should not be occupied more than 500 ms.
>     *   We suggest using a timer to check periodically, if users need to call os_delay_us or function
while , or function for in timer callback, please do not occupy CPU more than 15 ms.
>     *   Using non-OS SDK, please do not call any function defined with ICACHE_FLASH_ATTR in the
interrupt handler.