@Alkorin 2016-01-10
Ok, so I discard my idea to let a "safe area" at the end of the stream. 10µs is the time to send 8 bits... :)