@TerryE 2016-01-11
It's a guideline. Masking interrupts around an RGB set is less than 3x this, so probably won't cause problems and will  ensure that only complete RGB sets get shifted out. Allowing interrupts between sets and checking the clock delta around the enable - disable to detect for interrupt delivered and restart on interrupt detected is what I'd try.