@Alkorin 2016-01-17
@skycoders have fun, it works on my 300 leds strip.
Timings measured with my scope: 
0 => ~320ns 
1 => ~944ns
Period => 1.25us

It's a first draft, I've to clean a little bit :D